+++ db/migrate/20180215175758_remove_favorited_from_project_preferences.rb
@@ -1,0 +1,5 @@
+class RemoveFavoritedFromProjectPreferences < ActiveRecord::Migration[5.2]
+  def change
+    remove_column :project_preferences, :favorited, :boolean, null: false, default: false
+  end
+end